refactor(core): move EmbeddingProvider + companion types to julie-core
Move EmbeddingBackend, EmbeddingRuntimeStatus, DeviceInfo, and the
EmbeddingProvider trait into crates/julie-core/src/embeddings_contract.rs.
The src/embeddings/mod.rs keeps a pub use shim so every existing
crate::embeddings::* import path across ~20 prod + 15 test sites compiles
unchanged — no import-path edits anywhere in the main crate.

Gates:
  cargo check             ✓ (zero errors, zero new warnings)
  test_device_info_*      ✓ PASS (DeviceInfo::is_accelerated exercises moved type)
  test_embedding_runtime* ✓ PASS (EmbeddingRuntimeStatus exercises moved type)
  cargo build             ✓ (full workspace)
