+++ crates/julie-core/src/lib.rs
@@ -3,3 +3,5 @@
 //! This crate holds shared types and infrastructure that the top-level `julie`
 //! crate (and any future sibling crates) depend on. It must remain a true leaf:
 //! no references to `crate::handler`, `crate::tools`, or `crate::daemon`.
+
+pub mod embeddings_contract;
